refactor: split MCP server.rs into domain modules
Pure structural refactor — zero logic changes. server.rs (2,579 lines)
split into 5 files by domain responsibility:

  mod.rs   (158)  — entry, dispatch, thin router
  tools.rs (428)  — all 20 tool schema definitions
  scry.rs  (1140) — retrieval handlers + format helpers
  assay.rs (544)  — structural query handlers
  spec.rs  (367)  — spec lifecycle + schema introspection
